refact: TBTI, TEAM, USER 관련 model 패키지 분할 및 파일 분할
