+++ Tlog/app/src/main/java/com/tlog/api/TbtiApi.kt
@@ -1,7 +1,7 @@
 package com.tlog.api
 
 import com.tlog.data.model.response.base.BaseResponse
-import com.tlog.data.api.UpdateTbtiResponse
+import com.tlog.data.model.response.tbti.UpdateTbtiResponse
 import com.tlog.data.model.share.TbtiDescription
 import com.tlog.data.model.tbti.TbtiQuestion
 import retrofit2.http.GET
