--- Tlog/app/src/main/java/com/tlog/data/dto/UserProfileDto.kt
@@ -1,5 +1,0 @@
-package com.tlog.data.dto
-
-data class UserProfileDto(
-    val tbtiValue: String
-)
