+++ Tlog/app/src/main/java/com/tlog/data/model/request/auth/ProfileImageRequest.kt
@@ -1,0 +1,5 @@
+package com.tlog.data.model.request.auth
+
+data class ProfileImageRequest(
+    val imageUrl: String
+)
