+++ Tlog/app/src/main/java/com/tlog/data/model/request/auth/RegisterRequest.kt
@@ -1,10 +1,10 @@
 package com.tlog.data.model.request.auth
 
-import com.tlog.data.dto.UserProfileDto
+import com.tlog.data.model.request.auth.UserProfile
 
 data class RegisterRequest(
     val type: String,
     val accessToken: String,
-    val userProfile: UserProfileDto,
+    val userProfile: UserProfile,
     val preferTagIds: List<Int>
 )
