+++ Tlog/app/src/main/java/com/tlog/data/model/request/auth/UserProfile.kt
@@ -1,0 +1,5 @@
+package com.tlog.data.model.request.auth
+
+data class UserProfile(
+    val tbtiValue: String
+)
