+++ Tlog/app/src/main/java/com/tlog/data/model/request/team/CreateTeamRequest.kt
@@ -1,0 +1,7 @@
+package com.tlog.data.model.request.team
+
+data class CreateTeamRequest(
+    val name: String,
+    val creator: String,
+    val travelPlan: TravelPlanBody
+)
