+++ Tlog/app/src/main/java/com/tlog/data/model/request/team/JoinTeamRequest.kt
@@ -1,0 +1,6 @@
+package com.tlog.data.model.request.team
+
+data class JoinTeamRequest(
+    val inviteCode: String,
+    val userId: String
+)
