+++ Tlog/app/src/main/java/com/tlog/data/model/request/team/TravelPlanBody.kt
@@ -1,0 +1,11 @@
+package com.tlog.data.model.request.team
+
+data class TravelPlanBody(
+    val city: String,
+    val regionList: List<String>,
+    val hasPet: Boolean,
+    val hasTransport: Boolean,
+    val startDate: String,
+    val endDate: String,
+    val visitCountPerDay: Map<String, Int>
+)
