+++ Tlog/app/src/main/java/com/tlog/data/model/response/tbti/UpdateTbtiResponse.kt
@@ -1,0 +1,8 @@
+package com.tlog.data.model.response.tbti
+
+data class UpdateTbtiResponse(
+    val tbtiString: String,
+    val imageUrl: String,
+    val secondName: String,
+    val description: String
+)
