+++ Tlog/app/src/main/java/com/tlog/data/model/response/team/TeamCreateResponse.kt
@@ -1,0 +1,6 @@
+package com.tlog.data.model.response.team
+
+data class TeamCreateResponse(
+    val teamId: String,
+    val chatRoomId: Int
+)
