+++ Tlog/app/src/main/java/com/tlog/data/repository/MyPageRepository.kt
@@ -4,8 +4,8 @@ import android.util.Log
 import com.tlog.api.LoginApi
 import com.tlog.api.UserApi
 import com.tlog.data.model.response.base.BaseResponse
-import com.tlog.data.api.FeedbackRequest
-import com.tlog.data.api.ProfileImageRequest
+import com.tlog.data.model.request.auth.FeedbackRequest
+import com.tlog.data.model.request.auth.ProfileImageRequest
 import com.tlog.data.model.user.User
 import javax.inject.Inject
