+++ Tlog/app/src/main/java/com/tlog/data/repository/TbtiRepository.kt
@@ -2,7 +2,7 @@ package com.tlog.data.repository
 
 import com.tlog.api.TbtiApi
 import com.tlog.data.model.response.base.BaseResponse
-import com.tlog.data.api.UpdateTbtiResponse
+import com.tlog.data.model.response.tbti.UpdateTbtiResponse
 import com.tlog.data.model.share.TbtiDescription
 import com.tlog.data.model.tbti.TbtiQuestion
 import kotlinx.coroutines.Dispatchers
