+++ Tlog/app/src/main/java/com/tlog/data/repository/TeamRepository.kt
@@ -3,9 +3,9 @@ package com.tlog.data.repository
 import android.util.Log
 import com.tlog.api.TeamApi
 import com.tlog.data.model.response.base.BaseResponse
-import com.tlog.data.api.CreateTeamRequest
-import com.tlog.data.api.JoinTeamRequest
-import com.tlog.data.api.TeamCreateResponse
+import com.tlog.data.model.request.team.CreateTeamRequest
+import com.tlog.data.model.request.team.JoinTeamRequest
+import com.tlog.data.model.response.team.TeamCreateResponse
 import com.tlog.data.model.team.DetailTeam
 import com.tlog.data.model.team.Team
 import jakarta.inject.Inject
