+++ Tlog/app/src/main/java/com/tlog/viewmodel/beginning/ChooseMyTypeViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.mutableStateOf
 import com.tlog.data.local.TokenProvider
 import com.tlog.data.model.request.auth.FcmTokenRequest
 import com.tlog.data.model.request.auth.RegisterRequest
-import com.tlog.data.dto.UserProfileDto
+import com.tlog.data.model.request.auth.UserProfile
 import com.tlog.data.local.UserPreferences
 import com.tlog.data.repository.ChooseMyTypeRepository
 import com.tlog.ui.navigation.Screen
@@ -46,7 +46,7 @@ class ChooseMyTypeViewModel @Inject constructor(
                 val request = RegisterRequest(
                     type = socialType.toString(),
                     accessToken = socialAccessToken,
-                    userProfile = UserProfileDto(tbtiValue = tbtiValue),
+                    userProfile = UserProfile(tbtiValue = tbtiValue),
                     preferTagIds = _selected.value
                 )
