+++ Tlog/app/src/main/java/com/tlog/viewmodel/share/FeedbackViewModel.kt
@@ -2,7 +2,7 @@ package com.tlog.viewmodel.share
 
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
-import com.tlog.data.api.FeedbackRequest
+import com.tlog.data.model.request.auth.FeedbackRequest
 import com.tlog.data.repository.MyPageRepository
 import com.tlog.viewmodel.base.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
